pyproject.toml
[project]
name = "python-training"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "notebook>=7.4.7",
    "pandas>=2.3.3",
    "polars>=1.35.1",
    "python-decouple>=3.8",
    "python-dotenv>=1.2.1",
    "telethon>=1.41.2",
]
